pyproject.toml
[project]
name = "mlip-pipelines"
version = "0.1.0"
description = "automatic mlip builder"
authors = [
    { name = "tomoyuki tsuyama", email = "[tomoyuki.tsuyaman@gmail.com]" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ase>=3.27.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "F", "E", "W", "C90", "I", "N", "UP", "YTT", "ANN", "ASYNC", "S", "BLE",
    "B", "A", "C4", "DTZ", "T10", "EM", "ICN", "PIE", "T20", "PT", "RET",
    "SIM", "TID", "ARG", "PTH", "ERA", "PL", "TRY", "RUF"
]

ignore = [
    "ANN101", "ANN002", "ANN003", "ANN001", "ANN401", "ARG001", "ARG005",
    "PLR2004", "E501", "TRY003", "D", "ANN201", "N806", "PLC0415", "BLE001",
    "PT019", "RUF003", "ARG002", "RUF043", "RUF059", "PLR0913", "ANN202"
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # Allow assert in tests
"tutorials/UAT_AND_TUTORIAL.py" = ["N999"]

[tool.ruff.lint.mccabe]
max-complexity = 10  # Exceeding this prompts function splitting (prevent AI long functions)

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true  # Arguments starting with _ do not require type hints

[tool.pytest.ini_options]
addopts = "--cov=dev_src --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.mypy]
strict = true
ignore_missing_imports = true
